+++ test/__fixtures/test/integration/security/component-permission-removal/index.js
@@ -1,0 +1,14 @@
+class Component {
+  constructor() {}
+  static create() {
+    return new Component();
+  }
+
+  initialize = function($happn, callback) {
+    callback();
+  };
+   causeEmit($happn, eventKey, data, callback) {
+   $happn.emit(eventKey, data, null, callback);   
+  }
+}
+module.exports = Component;